% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators-pt.R
\name{adjusted.estimate.pt}
\alias{adjusted.estimate.pt}
\title{Adjusted estimator  accounting for prior test results.}
\usage{
adjusted.estimate.pt(
  n_r_pt,
  n_n,
  n_p,
  omega,
  beta,
  big_T,
  num_beta,
  den_omega,
  den_beta,
  q = 1
)
}
\arguments{
\item{n_r_pt}{Number of recent positives}

\item{n_n}{Number of negatives}

\item{n_p}{Number of positives}

\item{omega}{Mean duration of recent infection (MDRI) (in years, not days)}

\item{beta}{False recency rate (FRR)}

\item{big_T}{The \eqn{T^*} in the equation above}

\item{q}{The fraction of positives that were given recency tests, defaults to 1}
}
\description{
Adjusted estimator  accounting for prior test results.
}
