% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-study-sim.R
\name{assay.properties.est}
\alias{assay.properties.est}
\title{Use an assay df with recent (yes/no) and duration columns
to calculate the mean window period and MDRI of the assay}
\usage{
assay.properties.est(study, bigT, tau, last_point = TRUE)
}
\arguments{
\item{study}{Data frame with recent and durations variables}

\item{bigT}{The T^* time}

\item{tau}{The maximum time}
}
\value{
List of properties and their variances
}
\description{
Use an assay df with recent (yes/no) and duration columns
to calculate the mean window period and MDRI of the assay
}
