% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-study-sim.R
\name{assay.properties.nsim}
\alias{assay.properties.nsim}
\title{Function that simulates all the properties of
the recency assay, and returns all of the ones that we're interested in}
\usage{
assay.properties.nsim(n_sims, phi.func, bigT, tau, frr_mixture = NULL)
}
\description{
Function that simulates all the properties of
the recency assay, and returns all of the ones that we're interested in
}
