% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-study-sim.R
\name{assay.properties.sim}
\alias{assay.properties.sim}
\title{Function that simulates all the properties of
the recency assay, and returns all of the ones that we're interested in}
\usage{
assay.properties.sim(phi.func)
}
\description{
Function that simulates all the properties of
the recency assay, and returns all of the ones that we're interested in
}
