% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator.R
\name{avg.incidence}
\alias{avg.incidence}
\title{Compute the average incidence
over an interval.}
\usage{
avg.incidence(inc_function, ts, ...)
}
\description{
Compute the average incidence
over an interval.
}
