% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator.R
\name{c.infections}
\alias{c.infections}
\title{Function to produce infection times
from constant incidence.}
\usage{
\method{c}{infections}(e, t, p, lambda_0, rho = NA)
}
\description{
Function to produce infection times
from constant incidence.
}
