% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cephia}
\alias{cephia}
\title{CEPHIA Public Use Dataset}
\format{
\subsection{\code{cephia}}{

A data frame with 212831 rows and 38 columns.
}
}
\source{
\url{https://zenodo.org/record/4900634#.ZCW2HuzMI-R}
}
\usage{
cephia
}
\description{
Grebe, Eduard, Facente, Shelley N., Hampton,
Dylan, Marson, Kara, Hall, Jake, McKinney,
Elaine, Lebedeva, Mila, Parkin, Neil, Keating,
Sheila M., Kassanjee, Reshma, Pilcher,
Christopher D., Murphy, Gary, Busch,
Michael P., & Welte, Alex. (2021).
CEPHIA public use data (1.0) \link{Data set}.
Zenodo. https://doi.org/10.5281/zenodo.4900634
}
\keyword{datasets}
