% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator.R
\name{con.infections}
\alias{con.infections}
\title{Infection times function based on constant incidence.}
\usage{
con.infections(e, t, p, lambda_0, rho = NA)
}
\arguments{
\item{e}{A number between 0 and 1 (randomly generated), float or vector}

\item{t}{Time, float}

\item{p}{Constant prevalence}

\item{lambda_0}{Baseline incidence}
}
\value{
Infection time
}
\description{
Infection times function based on constant incidence.
}
\examples{
e <- runif(10)
con.infections(e, t=0, p=0.2, lambda_0=0.05)
}
