% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-rita-data.R
\name{createRitaCephia}
\alias{createRitaCephia}
\title{Create a dataset based on an algorithm and CEPHIA recency testing data.}
\usage{
createRitaCephia(
  assays,
  algorithm,
  subtype = NULL,
  ever_art = NULL,
  filepath = NULL
)
}
\arguments{
\item{assays}{A vector of assays to include}

\item{algorithm}{A function that defines the recency indicator
with arguments in the same order as the \code{assays} vector.
Arguments do not need to have the same name as \code{assays}.
E.g., if you have \code{assays = c("BED", "viral_load")},
you can have \code{algorithm = function(b, v) ...} where \code{b}
indicates BED and \code{v} indicates viral load.}

\item{subtype}{HIV subtypes to include (one of "A1", "B", "C", "D").
By default includes everyone.}

\item{ever_art}{Subset data to only those who have used ARTs or
have not. By default includes everyone.}

\item{filepath}{Optional full filepath to downloaded CEPHIA dataset (possibly obtained via the \code{download_cephia}).
If NULL, then uses CEPHIA version in this package.}
}
\value{
A data frame with the following columns:
\item{id}{ID column (there can be multiple measurements per individual)}
\item{ui}{infection duration in \strong{days} (note, if using with other functions in this package, may have to convert to years)}
\item{ri}{identified as recent based on algorithm}
}
\description{
Get data frame using CEPHIA data for a given recency
algorithm specified by the user. Optional subtype and ART exclusions.
}
\examples{
f <- function(b, l, v){
  ifelse((b > 1 & l < 3 & !is.na(v)), 1, 0)
}
createRitaCephia(assays=c("BED", "LAg-Sedia", "viral_load"),
             algorithm=f)
f <- function(l, v){
  v <- ifelse(l > 1.5, 0, v)
  return(
    ifelse((l <= 1.5) & (v > 1000), 1, 0)
  )
}
test <- createRitaCephia(assays=c("LAg-Sedia", "viral_load"), algorithm=f)
}
