% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi-matrix.R
\name{.matrix.phi}
\alias{.matrix.phi}
\title{Create a phi matrix of d x d based on draws from the phi.hat function}
\usage{
.matrix.phi(model, family, ts, ts_index)
}
\arguments{
\item{model}{A model object (either glm or gee)}

\item{ts_index}{Ts that should be used to form the matrix}
}
\description{
Create a phi matrix of d x d based on draws from the phi.hat function
}
