% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-rita-data.R
\name{downloadCephia}
\alias{downloadCephia}
\title{Download latest CEPHIA dataset}
\usage{
downloadCephia(path, timeout = 6000)
}
\arguments{
\item{path}{A folder to store the CEPHIA dataset}

\item{timeout}{The amount of time to allow R to download file}
}
\value{
A full filepath to the downloaded CEPHIA dataset
}
\description{
Downloads the latest CEPHIA dataset from concept DOI
10.5281/zenodo.4900633, and returns a local filepath where it can be found.
This function can be used to download the latest version, and use it in the
\code{\link[=getRitaOptions]{getRitaOptions()}} and \code{\link[=createRitaCephia]{createRitaCephia()}} functions.
}
