% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator.R
\name{e.infections}
\alias{e.infections}
\title{Function to produce infection times
from exponential incidence.}
\usage{
e.infections(e, t, p, lambda_0, rho)
}
\description{
Function to produce infection times
from exponential incidence.
}
