% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators-pt.R
\name{estEnhanced}
\alias{estEnhanced}
\title{Get enhanced estimate and variance (Bannick and Gao 2023+).
This is an estimator similar to the adjusted estimator, but including optional prior HIV test results.}
\usage{
estEnhanced(
  n_p,
  n,
  ptdf,
  beta,
  beta_var,
  big_T,
  phidat,
  use_geese,
  formula,
  family,
  plot_phi = TRUE,
  return_all = FALSE,
  ...
)
}
\arguments{
\item{n_p}{Number of positives}

\item{n}{Number of observations}

\item{ptdf}{Prior testing data frame. Only include those who are positive.}

\item{beta}{False recency rate (FRR)}

\item{beta_var}{Variance of the estimator for FRR (or 0 if FRR known)}

\item{big_T}{The \eqn{T^*} in the equation above}

\item{phidat}{Data to fit a modfunc model to. Must have column names ri
for recency indicator and ui for infection duration.
Can have more columns, e.g., if an id is needed for geese.}

\item{use_geese}{Indicator to fit a gee model using geese(), or a glm model with glm()}

\item{formula}{Formula for fitting the model to phidat.
Formula argument must take in only ui as the newdata.
For example, do not create a ui^2 variable. Use poly(ui, ...) function
to fit polynomial terms.}

\item{family}{Family argument for glm or gee}

\item{plot_phi}{Whether to plot the estimated phi function}

\item{return_all}{Return information to be used internally (not for general usage)}
}
\value{
Returns a list of the estimate and the variance, or more:
\item{est}{incidence estimate}
\item{var}{estimate of variance}
\item{omega}{estimate of MDRI using \code{phidat} and model specs}
\item{omega_var}{estimate of variance of MDRI using \code{phidat} and model specs}
\item{n_r}{number of recent infections identified based on recency assay alone, from \code{ptdf}}
\item{n_r_pt}{number of recent infections after applying enhanced algorithm using prior tests}
\item{q_eff}{number of HIV-pos individuals with recency assay results + prior tests / recency assay results}
\item{q}{fraction of HIV-pos individuals with recency assay results / HIV-pos individuals}
}
\description{
Get enhanced estimate and variance (Bannick and Gao 2023+).
This is an estimator similar to the adjusted estimator, but including optional prior HIV test results.
}
\details{
For example of usage, see the enhanced estimator \code{vignette("enhanced", package = "XSRecency")}
}
