% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{estSnapshot}
\alias{estSnapshot}
\title{Snapshot estimator and variance (Kaplan and Brookmeyer 1999)}
\usage{
estSnapshot(n_r, n_n, n_p, n, mu, mu_var, q = 1)
}
\arguments{
\item{n_r}{Number of recent positives}

\item{n_n}{Number of negatives}

\item{n_p}{Number of positives}

\item{n}{Total number of observations (\eqn{n_n + n_p = n})}

\item{mu}{Mean window period (in years, not days)}

\item{mu_var}{Variance of the estimator for mean window period.}

\item{q}{Fraction of positives that were given recency test, defaults to 1
If \eqn{\mu} is known, input 0.}
}
\value{
Returns a list of the estimate and the variance for \eqn{\lambda}.
}
\description{
\deqn{
    \hat{\lambda} = \frac{N_{rec}/q}{\hat{\mu} N_{neg}}
  }
where \eqn{\hat{\mu}} is an estimate of the mean window period.
}
\details{
\code{q} is an adjustment for the number of positives that are given recency tests.
}
\examples{
estSnapshot(n_r=2, n_n=50, n_p=10, n=60,
             mu=0.36, mu_var=0)
estSnapshot(n_r=c(2, 3), n_n=c(50, 48), n_p=c(10, 12), n=c(60, 60),
             mu=0.36, mu_var=0)
}
