% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator.R
\name{exn.incidence}
\alias{exn.incidence}
\title{Exponentially decreasing incidence function}
\usage{
exn.incidence(t, lambda_0, rho = 1)
}
\arguments{
\item{t}{Time, float or vector}

\item{lambda_0}{Incidence at time 0}

\item{rho}{Exponential decrease in incidence}
}
\value{
Incidence at time t
}
\description{
\deqn{
  \lambda(t) = \lambda_0 \exp(-\rho t)
}
}
\examples{
exn.incidence(0, lambda_0=0.05, rho=0.07)
exn.incidence(c(-1, 0, 1), lambda_0=0.05, rho=0.07)
}
