% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator.R
\name{exn.infections}
\alias{exn.infections}
\title{Infection times function based on exponential decreasing incidence.}
\usage{
exn.infections(e, t, p, lambda_0, rho)
}
\arguments{
\item{e}{A number between 0 and 1 (randomly generated), float or vector}

\item{t}{Time, float}

\item{p}{Constant prevalence}

\item{lambda_0}{Baseline incidence}

\item{rho}{Linearly decreasing incidence parameter}
}
\value{
Infection time
}
\description{
Infection times function based on exponential decreasing incidence.
}
\examples{
e <- runif(10)
exn.infections(e, t=0, p=0.2, lambda_0=0.05, rho=0.07)
}
