% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-study-sim.R
\name{fit.cubic}
\alias{fit.cubic}
\title{Fits a cubic model to indicators and durations
and returns the model object.}
\usage{
fit.cubic(recent, durations, id)
}
\description{
Fits a cubic model to indicators and durations
and returns the model object.
}
