% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-study-sim.R
\name{fit.model}
\alias{fit.model}
\title{Helper function for simulating many studies
that mimic the data from Duong et al. 2015.
Fits a GEE model with piece-wise linear
parameterization and Poisson marginal mean / variance.}
\usage{
fit.model(df, knot = 5)
}
\arguments{
\item{df}{Data frame with sample number column (samp),
days column, and num.samples column.}

\item{knot}{Knot location for piecewise-linear function}
}
\value{
List of start durations, numbers of samples, and model coefficients
}
\description{
Helper function for simulating many studies
that mimic the data from Duong et al. 2015.
Fits a GEE model with piece-wise linear
parameterization and Poisson marginal mean / variance.
}
