% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator.R
\name{generate.raw.data}
\alias{generate.raw.data}
\title{Generates raw data based on study parameters
that can be used later to simulate infection times}
\usage{
generate.raw.data(n_sims, n, prevalence)
}
\arguments{
\item{n_sims}{Number of simulations}

\item{n}{Number of observations per time point}

\item{prevalence}{Constant prevalence}
}
\description{
Generates raw data based on study parameters
that can be used later to simulate infection times
}
