% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi-functions.R
\name{get.gamma.params}
\alias{get.gamma.params}
\title{Gets the Gamma alpha and beta parameters
for a given mean window period and shadow}
\usage{
get.gamma.params(window, shadow)
}
\description{
Gets the Gamma alpha and beta parameters
for a given mean window period and shadow
}
