% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-rita-data.R
\name{get.rita.data}
\alias{get.rita.data}
\title{Get recency assay data for a phi dataset}
\usage{
get.rita.data(assay_name, threshold, subtype = NULL, ever_art = NULL)
}
\arguments{
\item{subtype}{Optional HIV subtypes as a vector (see `rita.subtype()` for options)}

\item{assay}{Recent infection testing algorithm to use (see `rita.subtype()` for options)}
}
\value{
A dataset with `ri` (recency indicator), `ui` (infection duration [in]) and `id` (participant ID)
}
\description{
Get recency assay data for a phi dataset
}
