% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-rita-data.R
\name{getRitaOptions}
\alias{getRitaOptions}
\title{See which recency assays are available by HIV subtype}
\usage{
getRitaOptions(filepath = NULL)
}
\arguments{
\item{filepath}{Optional full filepath to downloaded CEPHIA dataset (possibly obtained via the \code{downloadCephia} function).
If NULL, then uses CEPHIA version in this package.}
}
\description{
See which recency assays are available by HIV subtype
}
\examples{
getRitaOptions()
}
