% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi-functions.R
\name{getTestRecentFunc}
\alias{getTestRecentFunc}
\title{Approximate test-recent function with given window and shadow period}
\usage{
getTestRecentFunc(window, shadow)
}
\arguments{
\item{window}{Window period (in days)}

\item{shadow}{Shadow period (in days)}
}
\value{
A test-recent function of t
}
\description{
Generate a test-recent function that has a desired window and shadow period
using a gamma distribution parameters.
}
\examples{

phi <- getTestRecentFunc(window=200, shadow=191)
}
