% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi-functions.R
\name{getTestRecentFuncAdj}
\alias{getTestRecentFuncAdj}
\title{Generate a test-recent function that has a desired window and shadow period
that satisfies the shape assumption of the adjusted estimator.}
\usage{
getTestRecentFuncAdj(mdri, shadow, bigT, frr)
}
\arguments{
\item{mdri}{MDRI (in days)}

\item{shadow}{Shadow period (in days)}

\item{bigT}{The T that is the cutoff for recent infection}

\item{frr}{The desired false recency rate (FRR)}
}
\value{
A test-recent function of t
}
\description{
Generate a test-recent function that has a desired window and shadow period
that satisfies the shape assumption of the adjusted estimator.
}
\examples{

phi <- getTestRecentFuncAdj(mdri=120, shadow=100, bigT=2, frr=0.02)
}
