% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator.R
\name{incidence.lin}
\alias{incidence.lin}
\title{Linearly decreasing incidence function}
\usage{
incidence.lin(t, lambda_0, rho = 1)
}
\arguments{
\item{t}{Time, float or vector}

\item{lambda_0}{Incidence at time 0}

\item{rho}{Linear decrease in incidence}
}
\value{
Incidence at time t
}
\description{
\deqn{
  \lambda(t) = \lambda_0 - \rho t
}
}
\examples{
incidence.lin(0, lambda_0=0.05, rho=1e-3)
incidence.lin(c(-1, 0, 1), lambda_0=0.05, rho=1e-3)
}
