% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator.R
\name{infections.lincon}
\alias{infections.lincon}
\title{Infection times function based on constant incidence until bigT
then linearly increasing after. Compared to the other infection functions
we've switched to 1-e, which doesn't matter because e is uniform.}
\usage{
infections.lincon(e, t, p, lambda_0, rho, bigT)
}
\description{
Infection times function based on constant incidence until bigT
then linearly increasing after. Compared to the other infection functions
we've switched to 1-e, which doesn't matter because e is uniform.
}
