% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-study-sim.R
\name{integrate.phi}
\alias{integrate.phi}
\title{Estimate omega_T from a model object
Can instead estimate mu hat if you pass in a sufficiently
long follow_T parameter (like 9)}
\usage{
integrate.phi(model, follow_T = PHI_PARAMS$FOLLOW_T)
}
\description{
Estimate omega_T from a model object
Can instead estimate mu hat if you pass in a sufficiently
long follow_T parameter (like 9)
}
