% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi-functions.R
\name{integratePhi}
\alias{integratePhi}
\title{Integrate a phi function to calculate the MDRI or window period}
\usage{
integratePhi(phi.function, maxT)
}
\arguments{
\item{phi.function}{A function of t (years) which has range 0 to at least \code{maxT}}

\item{maxT}{The t to integrate to}
}
\description{
Integrate a phi function to calculate the MDRI or window period
}
