% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator.R
\name{l.infections}
\alias{l.infections}
\title{Function to produce infection times
from linear incidence.}
\usage{
l.infections(e, t, p, lambda_0, rho)
}
\description{
Function to produce infection times
from linear incidence.
}
