% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator.R
\name{lin.infections}
\alias{lin.infections}
\title{Infection times function based on linearly decreasing incidence.}
\usage{
lin.infections(e, t, p, lambda_0, rho)
}
\arguments{
\item{e}{A number between 0 and 1 (randomly generated), float or vector}

\item{t}{Time, float}

\item{p}{Constant prevalence}

\item{lambda_0}{Baseline incidence}

\item{rho}{Linearly decreasing incidence parameter}
}
\value{
Infection time
}
\description{
Infection times function based on linearly decreasing incidence.
}
\examples{
e <- runif(10)
lin.infections(e, t=0, p=0.2, lambda_0=0.05, rho=1e-3)
}
