% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator.R
\name{modify.pt.generator}
\alias{modify.pt.generator}
\title{Modifies the prior testing data for recall bias.}
\usage{
modify.pt.generator(
  t_noise = function(t) t,
  t_range = NULL,
  d_misrep = 0,
  p_misrep = 0,
  t_total_exclude = NULL
)
}
\arguments{
\item{t_noise}{Function to add noise to a single prior testing time.
Must be a function of t, the prior testing time.}

\item{t_range}{Range of times in which to keep prior testing times.}

\item{d_misrep}{Probability of given a positive prior test, individual
reports negative prior test.}

\item{p_misrep}{Probability of not reporting a prior test result,
among those with positive prior test.}
}
\description{
Modifies the prior testing data for recall bias.
}
\examples{
e.func <- function(e) infections.lin(e, t=0, p=0.29, lambda=0.032, rho=0.07)
params <- get.gamma.params(window=200/365.25, shadow=191/365.25)
phi.func <- function(t) 1-pgamma(t, shape = params[1], rate = params[2])
sim <- sim.screening.generator(prevalence=0.29, e.func=e.func, phi.func=phi.func)
df <- sim(100)
sim.pt <- sim.pt.generator(ptest.dist=function(u) runif(1, 0, 4),
                         ptest.prob=function(u) 0.5)
pt.df <- sim.pt(df[df$di == 1,])
t_noise <- function(t) max(0, t + rnorm(n=1, sd=0.5))
modify.pt <- modify.pt.generator(t_noise=t_noise, d_misrep=1, p_misrep=0, t_range=c(1, 2))
pt.df2 <- modify.pt(pt.df)

}
