% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi-functions.R
\name{phi.character.1}
\alias{phi.character.1}
\title{Characteristic function for phi, with a given
MDRI and FRR. If FRR = 0, then MDRI is the "window" period.}
\usage{
phi.character.1(t, mdri = 142, frr = 0.015)
}
\description{
Characteristic function for phi, with a given
MDRI and FRR. If FRR = 0, then MDRI is the "window" period.
}
