% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi-functions.R
\name{phi.character.2}
\alias{phi.character.2}
\title{Characteristic function for phi, with
increasing phi(t) in the tail
such that long-infected individuals have an increasing
probability of testing recent}
\usage{
phi.character.2(t, mdri = 142, frr = 0.015)
}
\description{
Characteristic function for phi, with
increasing phi(t) in the tail
such that long-infected individuals have an increasing
probability of testing recent
}
