% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-study-sim.R
\name{phi.hat}
\alias{phi.hat}
\title{Create phi probability predictions
based on a model object and d durations}
\usage{
phi.hat(d, model)
}
\description{
Create phi probability predictions
based on a model object and d durations
}
