% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator-old.R, R/data-generator.R
\name{sim.infection.times}
\alias{sim.infection.times}
\title{Function to simulate infection times from an arbitrary incidence
function and with constant prevalence.}
\usage{
sim.infection.times(p, inc.function, nsims = 1000, dt = 0.001)

sim.infection.times(p, inc.function, nsims = 1000, dt = 0.001)
}
\arguments{
\item{p}{Constant prevalence value}

\item{inc.function}{A function of time that returns incidence}

\item{nsims}{Number of infection times to simulate}

\item{dt}{The precision for numerical integration. May need to increase.}
}
\value{
A vector of infection durations

A vector of infection durations
}
\description{
Function to simulate infection times from an arbitrary incidence
function and with constant prevalence.

Function to simulate infection times from an arbitrary incidence
function and with constant prevalence.
}
\examples{
set.seed(1)
sim.infection.times(p=0.2, nsims=15,
                    inc.function=function(t) incidence.con(t, 0.5))
sim.infection.times(p=0.2, nsims=15,
                    inc.function=function(t) incidence.con(t, 0.05))
set.seed(1)
sim.infection.times(p=0.2, nsims=15,
                    inc.function=function(t) incidence.con(t, 0.5))
sim.infection.times(p=0.2, nsims=15,
                    inc.function=function(t) incidence.con(t, 0.05))
}
