% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator.R
\name{sim.screening.generator}
\alias{sim.screening.generator}
\title{Generates raw data of number positive, negative, and screened
based on study parameters that can be used later to simulate infection times.}
\usage{
sim.screening.generator(prevalence, e.func, phi.func)
}
\arguments{
\item{prevalence}{Constant prevalence}

\item{e.func}{Function to simulate infection time from a 0-1 value}

\item{n_sims}{Number of simulations}

\item{n}{Number of observations per time point}
}
\description{
Generates raw data of number positive, negative, and screened
based on study parameters that can be used later to simulate infection times.
}
\examples{
e.func <- function(e) infections.lin(e, t=0, p=0.29, lambda=0.032, rho=0.07)
params <- get.gamma.params(window=200/365.25, shadow=191/365.25)
phi.func <- function(t) 1-pgamma(t, shape = params[1], rate = params[2])
sim <- sim.screening.generator(prevalence=0.29, e.func=e.func, phi.func=phi.func)
sim(100)

}
