% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator.R
\name{simCrossSect}
\alias{simCrossSect}
\title{Simulates cross-sectional data based on epidemiological parameters.}
\usage{
simCrossSect(
  phi.func,
  incidence_type,
  prevalence,
  baseline_incidence,
  rho = NA,
  bigT = NA
)
}
\arguments{
\item{incidence_type}{Incidence type: constant, linear, exponential, or linear-constant (constant up to \code{bigT} argument, linear after)}

\item{prevalence}{Constant prevalence}

\item{baseline_incidence}{Baseline incidence}

\item{rho}{Parameter for linear or exponentially decreasing incidence}

\item{n_sims}{Number of simulations}

\item{n}{Number of observations per time point}

\item{p}{Constant prevalence}

\item{t}{Time t baseline (defaults to 0)}
}
\description{
These are the available incidence functions. \code{t} is relative to time 0.
So, to get incidence 2 years ago, \code{t = -2}.
}
\details{
Constant incidence: \deqn{\lambda(t) = \lambda_0 - \rho}
Linear incidence: \deqn{\lambda(t) = \lambda_0 - \rho t}
Exponential incidence: \deqn{\lambda(t) = \lambda_0 \exp(-\rho t)}
Constant-linear incidence: \deqn{\lambda(t) = \lambda_0 - \rho (t + T^*) I(-t > T^*)}
}
\examples{
phi.func <- getTestRecentFunc(window=200, shadow=191)
sim <- simCrossSect(phi.func=phi.func,
                    incidence_type="linear", prevalence=0.29, baseline_incidence=0.032, rho=0.07)
sim(10)

}
