% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-study-sim.R
\name{simExternal}
\alias{simExternal}
\title{Function to simulate studies
that mimic the infection duration data from Duong et al. 2015 (or another external dataset)}
\usage{
simExternal(nsims, phi.func = NULL, ext_df = NULL)
}
\arguments{
\item{nsims}{Number of study simulations}

\item{phi.func}{Optional recency test-positive function}

\item{ext_df}{A dataset with column "duration" and column for binary "recent" indicator}
}
\value{
List of data frames with an id, time, and recency indicator
}
\description{
Function to simulate studies
that mimic the infection duration data from Duong et al. 2015 (or another external dataset)
}
\examples{
set.seed(0)
study <- simExternal(nsims=1, phi.func=function(t) 1 - pgamma(t, 1, 1.5))
head(study)
}
