% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator.R
\name{simPriorTests}
\alias{simPriorTests}
\title{Generates prior testing data for positive subjects
based on a prior testing distribution and prior testing probability}
\usage{
simPriorTests(ptest.dist, ptest.prob, ptest.dist2 = NULL)
}
\arguments{
\item{ptest.dist}{A prior test result distribution function, which must have
be a function of \code{u} which is an
infection duration
though it can ignore the \code{u} argument inside the function.
An example is \code{rnorm(1, mean=u)}.}

\item{ptest.prob}{Probability of prior test result being available.
Can instead be a function of \code{u}, the infection duration.}
}
\description{
Generates prior testing data for positive subjects
based on a prior testing distribution and prior testing probability
}
\examples{
set.seed(101)
phi.func <- getTestRecentFunc(window=200, shadow=191)
sim <- simCrossSect(phi.func=phi.func,
                    incidence_type="linear", prevalence=0.29, baseline_incidence=0.032, rho=0.07)
df <- sim(100)
sim.pt <- simPriorTests(ptest.dist=function(u) runif(1, 0, 4),
                        ptest.prob=function(u) 0.5)
sim.pt(df[df$di == 1,])

}
