% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-study-sim.R
\name{simulate.beta}
\alias{simulate.beta}
\title{This simulates the false recency rate using
tail probabilities from the phi parameters
by sampling from a binomial with a sample size
given in the study.}
\usage{
\method{simulate}{beta}()
}
\description{
This simulates the false recency rate using
tail probabilities from the phi parameters
by sampling from a binomial with a sample size
given in the study.
}
