% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-study-sim.R
\name{simulate.nbeta}
\alias{simulate.nbeta}
\title{This simulates many false recency rates using
tail probabilities from the phi parameters
by sampling from a binomial with a sample size
given in the study.}
\usage{
simulate.nbeta(nsims, phi.func, minT, maxT, studies = NULL)
}
\description{
This simulates many false recency rates using
tail probabilities from the phi parameters
by sampling from a binomial with a sample size
given in the study.
}
