% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator.R
\name{simulate.recent}
\alias{simulate.recent}
\title{Simulate recency indicators based on the data, true phi function,
and the infection incidence function.}
\usage{
\method{simulate}{recent}(
  sim_data,
  infection.function,
  phi.func,
  baseline_incidence,
  prevalence,
  rho
)
}
\arguments{
\item{infection.function}{Function that simulates the infection time}

\item{data}{Outputs from the `generate.raw.data` function}

\item{model}{A `glm` model that was fit using data generated from
the same true.phi.function, but then fit with cubic polynomial}
}
\description{
Simulate recency indicators based on the data, true phi function,
and the infection incidence function.
}
