% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-study-sim.R
\name{simulate.studies}
\alias{simulate.studies}
\title{Function to simulate many studies
that mimic the data from Duong et al. 2015}
\usage{
simulate.studies(nsims, phi.func = NULL, ext_df = NULL)
}
\arguments{
\item{nsims}{Number of study simulations}

\item{phi.func}{Optional recency test-positive function}
}
\value{
List of data frames with an id, time, and recency indicator
}
\description{
Function to simulate many studies
that mimic the data from Duong et al. 2015
}
