% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-study-sim.R
\name{simulate.study}
\alias{simulate.study}
\title{Simulate infection duration and recency indicators
from the study.}
\usage{
\method{simulate}{study}(phi.func)
}
\description{
Simulate infection duration and recency indicators
from the study.
}
