% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-summary.R
\name{summarize.pt.generator}
\alias{summarize.pt.generator}
\title{Function that summarizes a dataset + prior test results}
\usage{
\method{summarize}{pt.generator}(
  bigT,
  dt = 1/365.25,
  use_geese = FALSE,
  formula,
  family,
  plot_phi = TRUE,
  ...
)
}
\arguments{
\item{bigT}{The T^* parameter for defining recent infection.}

\item{dt}{An integration step size. Should be no more than 1 day.}

\item{use_geese}{Indicator to fit a gee model using geese(), or a glm model with glm()}

\item{formula}{Formula for fitting the model to phidat.
Formula argument must take in only ui as the newdata.
For example, do not create a ui^2 variable. Use poly(ui, ...) function
to fit polynomial terms.}

\item{family}{Family argument for glm or gee}

\item{...}{Additional arguments to either glm or geese(..., data=phidat).}

\item{ptdf}{Data frame of prior testing data. These records should
only be from people who are HIV positive. They need to have
recency testing data as well as prior testing data.}

\item{n}{Required total sample size n if `ptdf` is used instead of `df`}

\item{n_p}{Required number of positives if `ptdf` is used instead of `df`}

\item{df}{Instead of passing ptdf, can pass a data frame with one row
for each person, regardless of HIV status, and NA for all
of the prior testing data for HIV negative people}

\item{phidat}{Data to fit a modfunc model to. Must have column names ri
for recency indicator and ui for infection duration.
Can have more columns, e.g., if an id is needed for geese.}
}
\description{
Function that summarizes a dataset + prior test results
}
