% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi-functions.R
\name{true.frr}
\alias{true.frr}
\title{Non-constant FRR true rate based on the phi function and the tail probability of 0.015}
\usage{
true.frr(phi.function, bigT, tau)
}
\description{
Non-constant FRR true rate based on the phi function and the tail probability of 0.015
}
