% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi-functions.R
\name{true.window.mdri}
\alias{true.window.mdri}
\title{When using the shadow period, need to numerically calculate the MDRI
the MDRI argument here is really mean window period, just calling it
MDRI for convenience}
\usage{
true.window.mdri(phi.function, maxT)
}
\description{
When using the shadow period, need to numerically calculate the MDRI
the MDRI argument here is really mean window period, just calling it
MDRI for convenience
}
