% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators-pt.R, R/estimators.R
\name{var.log.to.var}
\alias{var.log.to.var}
\title{Convert log variance to variance}
\usage{
var.log.to.var(estimate, variance)

var.log.to.var(estimate, variance)
}
\description{
Convert log variance to variance

Convert log variance to variance
}
