% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{variance}
\alias{variance}
\title{Adjusted variance computation, is the same for adjusted
and snapshot estimator, just pass in beta_sim=list(est=0, var=0).}
\usage{
variance(n_n, n_r, n_p, n, omega, omega_var, beta, beta_var, big_T, q = 1)
}
\description{
Adjusted variance computation, is the same for adjusted
and snapshot estimator, just pass in beta_sim=list(est=0, var=0).
}
