% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-study-sim.R
\name{assay.properties.nsim}
\alias{assay.properties.nsim}
\title{Function that simulates the mean window period, mean duration of recent
infection, and false recent rate (FRR) based on external study data.}
\usage{
assay.properties.nsim(
  n_sims,
  phi.func,
  bigT,
  tau,
  ext_FRR = FALSE,
  ext_df = NULL,
  max_FRR = NULL,
  last_point = FALSE
)
}
\arguments{
\item{n_sims}{Number of simulations}

\item{phi.func}{A test-recent positive function of t}

\item{bigT}{The time cutoff value designating true recent versus false recent}

\item{tau}{The maximum duration of infection where a subject
could have a false-positive for recent infection}

\item{ext_FRR}{A dataset with column "duration" and column for binary "recent" indicator}

\item{max_FRR}{The maximum duration allowed}

\item{last_point}{Integrate the mean window period to the last observed
duration in the dataset, rather than tau}
}
\value{
A list of estimated mean window period \eqn{\mu} and its variance,
  estimated MDRI \eqn{\Omega_{T^*}} and its variance, and
  estimated FRR \eqn{\beta_{T^*}} and its variance.
}
\description{
The external study data for mean window and MDRI is available here from
Duong et al. 2015:
# https://doi.org/10.1371/journal.pone.0114947.s001
}
\examples{
set.seed(0)
assay.properties.sim(n_sims=1, phi.func=function(t) 1 - pgamma(t, 1, 1.5),
                     bigT=2, tau=12)
}
