% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-study-sim.R
\name{assay.properties.sim}
\alias{assay.properties.sim}
\title{Function that simulates the mean window period, mean duration of recent
infection, and false recent rate (FRR) based on external study data.}
\usage{
assay.properties.sim(phi.func, bigT, tau, frr_mixture = NULL)
}
\arguments{
\item{phi.func}{A test-recent positive function of t}

\item{bigT}{The time cutoff value designating true recent versus false recent}

\item{tau}{The maximum duration of infection where a subject
could have a false-positive for recent infection}

\item{frr_mixture}{Vector of two values.
To simulate data for FRR, we need a sample of long infected
subjects. If \code{frr_mixture} is \code{NULL}, then the sample of infection times is
\eqn{Uniform(2, 12)}. If \code{frr_mixture} is not \code{NULL},
then the sample is an equal mixture of \eqn{Uniform(2, 12)}
and \eqn{Uniform(x[1], x[2])} where \eqn{x} is \code{frr_mixture}.}
}
\value{
A list of estimated mean window period \eqn{\mu} and its variance,
  estimated MDRI \eqn{\Omega_{T^*}} and its variance, and
  estimated FRR \eqn{\beta_{T^*}} and its variance.
}
\description{
Function that simulates the mean window period, mean duration of recent
infection, and false recent rate (FRR) based on external study data.
}
\examples{
set.seed(0)
assay.properties.sim(phi.func=function(t) 1 - pgamma(t, 1, 1.5),
                     bigT=2, tau=12)
set.seed(0)
assay.properties.sim(phi.func=function(t) 1 - pgamma(t, 1, 1.5),
                     bigT=2, tau=12, frr_mixture=c(5, 8))
}
