% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator.R
\name{generate.data}
\alias{generate.data}
\title{Create simulations of data generated with past}
\usage{
generate.data(
  n_sims,
  n,
  infection.function,
  phi.func,
  baseline_incidence,
  prevalence,
  rho
)
}
\arguments{
\item{n_sims}{Number of simulations}

\item{n}{Number of subjects screened}

\item{infection.function}{Function that simulates the infection time}

\item{phi.func}{Test positive function for recency assay}

\item{baseline_incidence}{Baseline incidence value (time 0)}

\item{prevalence}{Constant prevalence value}

\item{rho}{A parameter to the `infection.function` for how quickly incidence changes}
}
\description{
Create simulations of data generated with past
}
