% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator.R
\name{generate.raw.data}
\alias{generate.raw.data}
\title{Generates raw data of number positive, negative, and screened
based on study parameters
that can be used later to simulate infection times.}
\usage{
generate.raw.data(n_sims, n, prevalence, times = c(0))
}
\arguments{
\item{n_sims}{Number of simulations}

\item{n}{Number of observations per time point}

\item{prevalence}{Constant prevalence}

\item{times}{Times at which enrollment happens}
}
\description{
Generates raw data of number positive, negative, and screened
based on study parameters
that can be used later to simulate infection times.
}
