% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{get.adjusted}
\alias{get.adjusted}
\title{Get adjusted estimate and variance.}
\usage{
get.adjusted(n_r, n_n, n_p, n, omega, omega_var, beta, beta_var, big_T)
}
\description{
Get adjusted estimate and variance.
}
