% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-study-data.R
\docType{data}
\name{PHI_PARAMS}
\alias{PHI_PARAMS}
\title{External study data parameters for estimating
omega, mu, and beta through simulation.}
\format{
An object of class \code{list} of length 9.
}
\usage{
PHI_PARAMS
}
\description{
External study data parameters for estimating
omega, mu, and beta through simulation.
}
\keyword{datasets}
