% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator.R
\name{simulate.recent}
\alias{simulate.recent}
\title{Simulate recency indicators based on the data, true phi function,
and the infection incidence function.}
\usage{
\method{simulate}{recent}(
  sim_data,
  infection.function = NULL,
  phi.func,
  baseline_incidence,
  prevalence,
  rho,
  summarize = TRUE,
  incidence.function = NULL,
  ...
)
}
\arguments{
\item{sim_data}{Outputs from the `generate.raw.data` function}

\item{infection.function}{Function that simulates the infection time}

\item{phi.func}{Test positive function for recency assay}

\item{baseline_incidence}{Baseline incidence value (time 0)}

\item{prevalence}{Constant prevalence value}

\item{rho}{A parameter to the `infection.function` for how quickly incidence changes}

\item{incidence.function}{An incidence function in place of `infection.function`.
Cannot pass both arguments.}

\item{...}{Additional arguments to the `simulate.infection.times` function if you're
passing an `incidence.function` instead of an `infection.function`.}
}
\description{
Simulate recency indicators based on the data, true phi function,
and the infection incidence function.
}
