% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-study-sim.R
\name{simulate.study}
\alias{simulate.study}
\title{Function to simulate one study
that mimics the data from Duong et al. 2015}
\usage{
\method{simulate}{study}(days, num.samples, coefs, knot = 5, phi.func = NULL)
}
\arguments{
\item{days}{Duration on date of first sample}

\item{num.samples}{Distribution of number of samples}

\item{coefs}{Coefficients from the piecewise linear model}

\item{knot}{Knot location for piecewise linear model}

\item{phi.func}{Optional recency test-positive function}
}
\value{
List of data frames with an id, time, and recency indicator
}
\description{
Function to simulate one study
that mimics the data from Duong et al. 2015
}
